+++ src/components/AddCardForm/index.stories.jsx
@@ -1,0 +1,10 @@
+import AddCardForm from '.';
+
+export default {
+  title: 'AddCardForm',
+  component: AddCardForm,
+};
+
+const Template = (args) => <AddCardForm {...args} />;
+
+export const Example = Template.bind({});
