+++ src/components/AddCardForm/index.styled.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const Container = styled.form`
+  display: flex;
+  flex-direction: column;
+  position: relative;
+`;
+
+export const ButtonContainer = styled.div`
+  position: absolute;
+  right: 20px;
+  bottom: 0;
+`;
