--- src/pages/AddCardPage/index.stories.jsx
@@ -1,10 +1,0 @@
-import AddCardPage from '.';
-
-export default {
-  title: 'AddCardPage',
-  component: AddCardPage,
-};
-
-const Template = (args) => <AddCardPage {...args} />;
-
-export const Example = Template.bind({});
